Jenkins Test + Coverage for frontend
The Jenkins pipeline is updated to report test and coverage report for the frontend. The coverage reports of the front and back ends are merged into one using a separate Node.js based script. The pipeline can be made faster to execute if the Jenkins Agent allocated for the frontend is kept alive for the merge stage. But, in future we may have other parts/services of the project that also produce such reports. As the merge step will remain to be one of the last steps, for best parallelism, the frontend resource is set free as soon as its done.